+++ extensions/iam/oauth2/oauth2-core/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/core/Oauth2Extension.java
@@ -20,11 +20,10 @@ import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.crypto.ECDSASigner;
 import com.nimbusds.jose.crypto.RSASSASigner;
 import okhttp3.OkHttpClient;
-import org.eclipse.dataspaceconnector.iam.oauth2.core.impl.DefaultJwtDecorator;
-import org.eclipse.dataspaceconnector.iam.oauth2.core.impl.IdentityProviderKeyResolver;
-import org.eclipse.dataspaceconnector.iam.oauth2.core.impl.JwtDecoratorRegistryImpl;
-import org.eclipse.dataspaceconnector.iam.oauth2.core.impl.Oauth2Configuration;
-import org.eclipse.dataspaceconnector.iam.oauth2.core.impl.Oauth2ServiceImpl;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.identity.IdentityProviderKeyResolver;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.identity.Oauth2ServiceImpl;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.jwt.DefaultJwtDecorator;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.jwt.JwtDecoratorRegistryImpl;
 import org.eclipse.dataspaceconnector.iam.oauth2.spi.JwtDecoratorRegistry;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
