+++ extensions/iam/oauth2/oauth2-core/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/core/identity/IdentityProviderKeyResolver.java
@@ -12,12 +12,14 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.iam.oauth2.core.impl;
+package org.eclipse.dataspaceconnector.iam.oauth2.core.identity;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.jwt.JwkKey;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.jwt.JwkKeys;
 import org.eclipse.dataspaceconnector.iam.oauth2.spi.PublicKeyResolver;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.types.TypeManager;
